Port Old Codebase
While working on a previous project, we created a datatype that was
ostensibly dynamic. One of the requirements was we should be able to
store multiple distinct datataypes and we did not know until the
execution phase what the datatype we would receive.

Through experimentation I discovered that would could store an
arbitrarilty large amount of distinct datatypes with not much
modification to the base code we developed.

The base code was forked from Postgres 11's JsonB codebase. Using this
we were able to add mutliple number types and other custom datatypes.

The goal of this project is to extract the old codebase to its own extension
and allow users to take advantage of this dynamic typing system in Postgres
without use of the rest of the project.

The extension will seemlessy support all builtin datatypes with little to
no effort and allow user to add other datatypes from other extensions or created
on their own with as little effort and work as possible.